+++ test/integration/policyfiles/default_test.rb
@@ -9,21 +9,22 @@ describe command('chef') do
   it { should exist }
 end
 
-describe command('CHEF_LICENSE="accept-no-persist" chef show-policy -c /etc/opscode/managed/config.rb') do
-  its ('stdout') { should match /\* _default:  bea04861be$/ }
-  its ('stdout') { should match /\* _default:  d99228eafe$/ }
-  its ('stdout') { should match /\* home:      3e28786370$/ }
+describe command('CHEF_LICENSE="accept-no-persist" chef show-policy -c /etc/opscode/managed/chef_managed_org/config.rb') do
+  its ('stdout') { should match /\* base:\s*bea04861be$/ }
+  its ('stdout') { should match /\* beaglebone:\s*d99228eafe$/ }
+  its ('stdout') { should match /\* macbookpro:\s*3e28786370$/ }
 end
 
-describe command('CHEF_LICENSE="accept-no-persist" chef show-policy base -c /etc/opscode/managed/config.rb') do
-  its ('stdout') { should match /\* _default:  bea04861be$/ }
+describe command('CHEF_LICENSE="accept-no-persist" chef show-policy base -c /etc/opscode/managed/chef_managed_org/config.rb') do
+  its ('stdout') { should match /\* base:\s*bea04861be$/ }
 end
 
-describe command('CHEF_LICENSE="accept-no-persist" chef show-policy beaglebone -c /etc/opscode/managed/config.rb') do
-  its ('stdout') { should match /\* _default:  d99228eafe$/ }
+describe command('CHEF_LICENSE="accept-no-persist" chef show-policy beaglebone -c /etc/opscode/managed/chef_managed_org/config.rb') do
+  its ('stdout') { should match /\* beaglebone:\s*d99228eafe$/ }
 end
 
-describe command('CHEF_LICENSE="accept-no-persist" chef show-policy macbookpro -c /etc/opscode/managed/config.rb') do
-  its ('stdout') { should match /\* home:      3e28786370$/ }
-  its ('stdout') { should match /\* _default:  \*NOT APPLIED\*$/ }
+describe command('CHEF_LICENSE="accept-no-persist" chef show-policy macbookpro -c /etc/opscode/managed/chef_managed_org/config.rb') do
+  its ('stdout') { should match /\* macbookpro:\s*3e28786370$/ }
+  its ('stdout') { should match /\* base:\s*\*NOT APPLIED\*$/ }
+  its ('stdout') { should match /\* beaglebone:\s*\*NOT APPLIED\*$/ }
 end